requirements.txt
# Core ML/DL
torch>=2.0.0
numpy>=1.24.0

# Tokenizer
tokenizers>=0.15.0
sentencepiece>=0.1.99

# Training utilities
wandb>=0.15.0
tqdm>=4.66.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2

# API & Serving
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.0
loguru>=0.7.0

# Optional: Flash Attention (requires separate installation)
# flash-attn>=2.0.0

# Optional: For distributed training
# deepspeed>=0.12.0

# AWS SageMaker
sagemaker>=2.200.0
boto3>=1.34.0

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
